+++ CHANGES.md
@@ -22,6 +22,7 @@
 
 * [SSHD-1193](https://issues.apache.org/jira/browse/SSHD-1193) Provide a more user-friendly text in case disconnecting due to timeout(s).
 * [SSHD-1196](https://issues.apache.org/jira/browse/SSHD-1196) Provide configurable support for SFTP output stream chunking behavior.
+* [SSHD-1226](https://issues.apache.org/jira/browse/SSHD-1226) Using literal constants for SFTP attributes key names
 
 ## Behavioral changes and enhancements